backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
shap==0.43.0
python-multipart==0.0.6
web3==6.11.3
py-solc-x==2.0.2
eth-account==0.10.0
